feat(mining): add mining services and orchestrators
Implement 11 mining services (creation, rule creation, slot buying/switching/calculation/selling/taking tokens, machine/slot queries, global rule query/creation) and 4 orchestrators (creation, slot taking token, switching, selling). Add MiningEngine DI registration, mining DbSets to ArkWalletDbContext and register services/orchestrators in StartProgram.

+++ ArkWallet/Domain/Engines/MiningEngine.cs
@@ -78,7 +78,7 @@ internal class MiningEngine
 
         var elapsed = (now - start.Value).TotalMinutes;
         var total = (end.Value - start.Value).TotalMinutes;
-        var percent = elapsed / total * 100m;
+        var percent = (decimal)(elapsed / total) * 100m;
         return Math.Clamp((decimal)percent, 0m, 100m);
     }
